LPS-178898 Extract ScreenNavigationEntry methods from DispatchLogScreenNavigationCategory to DispatchLogScreenNavigationEntry and extends DispatchLogScreenNavigationCategory in it
