+++ modules/apps/dispatch/dispatch-web/src/main/java/com/liferay/dispatch/web/internal/frontend/taglib/servlet/taglib/DispatchLogScreenNavigationCategory.java
@@ -15,27 +15,12 @@
 package com.liferay.dispatch.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.dispatch.constants.DispatchConstants;
-import com.liferay.dispatch.model.DispatchTrigger;
-import com.liferay.dispatch.service.DispatchLogService;
-import com.liferay.dispatch.web.internal.display.context.DispatchLogDisplayContext;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.util.JavaConstants;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
 import java.util.ResourceBundle;
-
-import javax.portlet.RenderRequest;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -44,15 +29,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=20",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=20",
+	service = ScreenNavigationCategory.class
 )
 public class DispatchLogScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<DispatchTrigger> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -60,16 +41,11 @@ public class DispatchLogScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return getCategoryKey();
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(resourceBundle, getCategoryKey());
+		return language.get(resourceBundle, getCategoryKey());
 	}
 
 	@Override
@@ -77,43 +53,7 @@ public class DispatchLogScreenNavigationCategory
 		return DispatchConstants.SCREEN_NAVIGATION_KEY_DISPATCH_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, DispatchTrigger dispatchTrigger) {
-		if (dispatchTrigger == null) {
-			return false;
-		}
-
-		return true;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		RenderRequest renderRequest =
-			(RenderRequest)httpServletRequest.getAttribute(
-				JavaConstants.JAVAX_PORTLET_REQUEST);
-
-		DispatchLogDisplayContext dispatchLogDisplayContext =
-			new DispatchLogDisplayContext(_dispatchLogService, renderRequest);
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT, dispatchLogDisplayContext);
-
-		_jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/trigger/dispatch_trigger_logs.jsp");
-	}
-
 	@Reference
-	private DispatchLogService _dispatchLogService;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
+	protected Language language;
 
 }
